// angular-frontend/src/app/models/user.model.ts
export interface User {
    group: string;
    cat: string;
    permissions: string[];
    studies: string[];
    matnr: string;
    name: string;
    surname: string;
    password: boolean;
    position: string;
    title_pre: string;
    title_aft: string;
    email: string;
    room: string;
    dial: string;
    phone: string;
    fax: string;
    mobile: string;
    website: string;
}